requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jinja2==3.1.4
mako==1.3.5
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.2
urllib3==2.2.2
uvicorn[standard]==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
